Outsource get_type_token and get_process_token functions to utils.py
