+++ experiments/eventgen/wrappers.py
@@ -6,31 +6,7 @@ from gatr.interface import embed_vector, extract_vector
 from gatr.layers import EquiLinear, GeometricBilinear, ScalarGatedNonlinearity
 from experiments.tagging.embedding import get_spurion
 from experiments.eventgen.cfm import EventCFM
-
-
-def get_type_token(x_ref, type_token_channels):
-    # embed type_token
-    type_token_raw = torch.arange(
-        x_ref.shape[-2], device=x_ref.device, dtype=torch.long
-    )
-    type_token = nn.functional.one_hot(type_token_raw, num_classes=type_token_channels)
-    type_token = type_token.expand(*x_ref.shape[:-1], type_token_channels)
-    return type_token
-
-
-def get_process_token(x_ref, ijet, process_token_channels):
-    # embed process_token
-    process_token_raw = torch.tensor([ijet], device=x_ref.device, dtype=torch.long)
-    process_token = nn.functional.one_hot(
-        process_token_raw, num_classes=process_token_channels
-    ).squeeze()
-    process_token = process_token.unsqueeze(0).expand(
-        x_ref.shape[1], process_token_channels
-    )
-    process_token = process_token.unsqueeze(0).expand(
-        x_ref.shape[0], x_ref.shape[1], process_token_channels
-    )
-    return process_token
+from experiments.eventgen.utils import get_type_token, get_process_token
 
 
 class MLPCFM(EventCFM):
